reorganized tests for series API
